fix: Move Redis backing to subpath export to prevent forced ioredis import
Fixes #10

This change addresses the issue where importing the main package would
automatically import `ioredis`, even for users who don't use Redis backing.
Since `ioredis` is marked as an optional peer dependency, this caused
runtime failures for consumers who only needed core functionality.

## Changes:

- Added subpath export `effect-distributed-lock/redis` for Redis functionality
- Created new `src/redis.ts` entry point for Redis backing
- Removed `RedisBacking` export from main `src/index.ts`
- Updated `package.json` with new subpath export configuration
- Updated README and examples to use new import path:
  - `import { RedisBacking } from "effect-distributed-lock/redis"`

## Breaking Change:

Users now need to import Redis backing from a separate path:

```typescript
// Before
import { DistributedSemaphore, RedisBacking } from "effect-distributed-lock";

// After
import { DistributedSemaphore } from "effect-distributed-lock";
import { RedisBacking } from "effect-distributed-lock/redis";
```

This ensures that `ioredis` is only imported when explicitly needed,
allowing users without Redis to use the core library without issues.
