+++ examples/push.ts
@@ -13,7 +13,8 @@
  */
 import { Console, Duration, Effect, Schedule } from "effect";
 import Redis from "ioredis";
-import { DistributedSemaphore, RedisBacking } from "../src/index.ts";
+import { DistributedSemaphore } from "../src/index.ts";
+import { RedisBacking } from "../src/redis.ts";
 
 const redis = new Redis(process.env.REDIS_URL ?? "redis://localhost:6379");
